requirements/base.txt
flask
mysql-connector-python
opencv-python
numpy
flask-cors